Chore: 변수명과 자료형 수정

+++ TogetUp/TogetUp/Screen/Login/Model/LoginModel.swift
@@ -32,7 +32,7 @@ struct LoginResult: Codable {
 
 struct UserStatus: Codable {
     let level: Int
-    let expPercentage: Int
+    let expPercentage: Double
     let point: Int
 }
